+++ lib/load_file.rb
@@ -8,11 +8,12 @@ module LoadFile
   #
   # @param [String, Pathname, File] path to load the file from
   # @param [String, Symbol] constant name to load into
+  # @param namespace [Object] namespace to find/load the constant, defaults to Object
   # @return [Hash] loaded file content
   # @return [NilClass] nil when file not exists
-  def self.load(file:, constant:)
+  def self.load(file:, constant:, namespace: Object)
     ignore_file_not_exists do
-      loader = Loader.new(file, constant)
+      loader = Loader.new(file, constant, namespace: namespace)
       loader.set_constant
     end
   end
@@ -21,10 +22,11 @@ module LoadFile
   #
   # @param [String, Pathname, File] path to load the file from
   # @param [String, Symbol] constant name to load into
+  # @param namespace [Object] namespace to find/load the constant, defaults to Object
   # @return [Hash] loaded file content, raises an error when file not exists
   # @return [NilClass] nil when file not exists
-  def self.load!(file:, constant:)
-    loader = Loader.new(file, constant)
+  def self.load!(file:, constant:, namespace: Object)
+    loader = Loader.new(file, constant, namespace: namespace)
     loader.set_constant
   end
 
@@ -35,8 +37,8 @@ module LoadFile
   # @param [String, Symbol] constant name to load into
   # @return [Hash] last loaded file content
   # @return [NilClass] nil when last file not exists
-  def self.load_files(files:, constant:)
-    files.each { |file| load(file: file, constant: constant) }
+  def self.load_files(files:, constant:, namespace: Object)
+    files.each { |file| load(file: file, constant: constant, namespace: namespace) }
   end
 
   # Loads files into constant.
@@ -44,8 +46,8 @@ module LoadFile
   # @param [Array<String>] list of files to load
   # @param [String, Symbol] constant name to load into
   # @return [Hash] last loaded file content, raises an error when any file not exists
-  def self.load_files!(files:, constant:)
-    files.each { |file| load!(file: file, constant: constant) }
+  def self.load_files!(files:, constant:, namespace: Object)
+    files.each { |file| load!(file: file, constant: constant, namespace: namespace) }
   end
 
   # Overload a `file` into `constant`.
@@ -55,9 +57,9 @@ module LoadFile
   # @param [String, Symbol] constant name to overload into
   # @return [Hash] overloaded file content
   # @return [NilClass] nil when file not exists
-  def self.overload(file:, constant:)
+  def self.overload(file:, constant:, namespace: Object)
     ignore_file_not_exists do
-      reader = Loader.new(file, constant)
+      reader = Loader.new(file, constant, namespace: namespace)
       reader.set_constant!
     end
   end
@@ -69,8 +71,8 @@ module LoadFile
   # @param [String, Symbol] constant name to overload into
   # @return [Hash] overloaded file content, raises an error when file not exists
   # @return [NilClass] nil when file not exists
-  def self.overload!(file:, constant:)
-    reader = Loader.new(file, constant)
+  def self.overload!(file:, constant:, namespace: Object)
+    reader = Loader.new(file, constant, namespace: namespace)
     reader.set_constant!
   end
 
@@ -82,8 +84,8 @@ module LoadFile
   # @param [String, Symbol] constant name to overload into
   # @return [Hash] last overloaded file content
   # @return [NilClass] nil when last file not exists
-  def self.overload_files(files:, constant:)
-    files.each { |file| overload(file: file, constant: constant) }
+  def self.overload_files(files:, constant:, namespace: Object)
+    files.each { |file| overload(file: file, constant: constant, namespace: namespace) }
   end
 
   # Overload files into constant.
@@ -93,8 +95,8 @@ module LoadFile
   # @param [Array<String>] list of files to overload
   # @param [String, Symbol] constant name to overload into
   # @return [Hash] last overloaded file content, raises an error when any file not exists
-  def self.overload_files!(files:, constant:)
-    files.each { |file| overload!(file: file, constant: constant) }
+  def self.overload_files!(files:, constant:, namespace: Object)
+    files.each { |file| overload!(file: file, constant: constant, namespace: namespace) }
   end
 
   # @private